Added support for Xamarin.iOS 64 bit (Unified API). Requires Xamarin.iOS 8.6.0.0 or newer. (iOS, Xamarin.Forms) Removed support for MonoTouch. Added UseAnimations hint for navigation, it allows to animate view during navigation (iOS, Xamarin.Forms). Added ICollectionViewManager and IContentViewManager for Android platform. Added BindingServiceProvider.BindingPathFactory static property that allows to create an instance of IBindingPath for the specified string. Updated nuspec files. Minor changes.

+++ Platforms/MugenMvvmToolkit.Touch/Binding/Infrastructure/CollectionViewSourceBase.cs
@@ -19,13 +19,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Foundation;
 using JetBrains.Annotations;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
 using MugenMvvmToolkit.Binding.Interfaces;
 using MugenMvvmToolkit.Binding.Interfaces.Models;
 using MugenMvvmToolkit.Binding.Modules;
 using MugenMvvmToolkit.Interfaces.Models;
+using UIKit;
 
 namespace MugenMvvmToolkit.Binding.Infrastructure
 {
@@ -264,7 +264,7 @@ namespace MugenMvvmToolkit.Binding.Infrastructure
                 PlatformDataBindingModule.CollectionViewCellHighlightedMember.Raise(cell, EventArgs.Empty);
         }
 
-        public override int NumberOfSections(UICollectionView collectionView)
+        public override nint NumberOfSections(UICollectionView collectionView)
         {
             return 1;
         }
